Move pipeline tests into their own module.
